Move buildComponents into main package #333
