--- packages/buildcomponents/.gitignore
@@ -1,5 +1,0 @@
-# Ignore Gradle project-specific cache directory
-.gradle
-
-# Ignore Gradle build output directory
-build
