--- packages/buildcomponents/.mocharc
@@ -1,4 +1,0 @@
-{
-  "require": " @babel/register"
-}
-cd pa
