--- packages/buildcomponents/.npmignore
@@ -1,6 +1,0 @@
-**/
-*
-*.*
-
-# Publish only these:
-!webpack.config.js
