--- packages/buildcomponents/CHANGELOG.md
@@ -1,43 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-## [1.0.6](https://github.com/enonic/react4xp-npm/compare/react4xp-build-components@1.0.5...react4xp-build-components@1.0.6) (2020-07-30)
-
-**Note:** Version bump only for package react4xp-build-components
-
-## [1.0.5](https://github.com/enonic/react4xp-npm/compare/react4xp-build-components@1.0.4...react4xp-build-components@1.0.5) (2020-07-30)
-
-**Note:** Version bump only for package react4xp-build-components
-
-## [1.0.4](https://github.com/enonic/react4xp-npm/compare/react4xp-build-components@1.0.3...react4xp-build-components@1.0.4) (2020-07-29)
-
-**Note:** Version bump only for package react4xp-build-components
-
-## [1.0.3](https://github.com/enonic/react4xp-npm/compare/react4xp-build-components@1.0.1...react4xp-build-components@1.0.3) (2020-07-29)
-
-**Note:** Version bump only for package react4xp-build-components
-
-## [1.0.2](https://github.com/enonic/react4xp-npm/compare/react4xp-build-components@1.0.1...react4xp-build-components@1.0.2) (2020-07-29)
-
-**Note:** Version bump only for package react4xp-build-components
-
-## [1.0.1](https://github.com/enonic/react4xp-npm/compare/react4xp-build-components@1.0.0...react4xp-build-components@1.0.1) (2020-07-29)
-
-**Note:** Version bump only for package react4xp-build-components
-
-## [0.7.7](https://github.com/enonic/react4xp-npm/compare/react4xp-build-components@0.7.6...react4xp-build-components@0.7.7) (2020-04-16)
-
-**Note:** Version bump only for package react4xp-build-components
-
-## [0.7.6](https://github.com/enonic/react4xp-npm/compare/react4xp-build-components@0.7.5...react4xp-build-components@0.7.6) (2020-03-27)
-
-**Note:** Version bump only for package react4xp-build-components
-
-## 0.7.5 (2020-03-21)
-
-### Bug Fixes
-
-* constants - published 0.8.2 because fixed
-  prepublish ([685a235](https://github.com/enonic/react4xp-npm/commit/685a23590c816a14ea74f387d09a0a0bffd5c2d8))
