--- packages/buildcomponents/__tests__/buildcomponents.test.js
@@ -1,7 +1,0 @@
-'use strict';
-
-const buildcomponents = require('..');
-
-describe('buildcomponents', () => {
-    it('needs tests');
-});
