--- packages/buildcomponents/gradle.properties
@@ -1,3 +1,0 @@
-xpVersion="7.0.0"
-org.gradle.parallel=true
-org.gradle.caching=true
