--- packages/buildcomponents/package.json
@@ -1,57 +1,0 @@
-{
-  "name": "react4xp-build-components",
-  "version": "2.0.1-SNAPSHOT",
-  "description": "React4xp helper: basic webpack compilation of user-added React components in a React4xp structure",
-  "main": "webpack.config.js",
-  "files": [
-    "entriesandchunks.js",
-    "webpack.config.js",
-    "README.md",
-    "LICENSE"
-  ],
-  "scripts": {
-    "test:main": "echo \"(Placeholder necessary for task dependencies)\"",
-    "test:entriesAndChunks": "mocha --require @babel/register --require @babel/polyfill --require test/test_helper.js \"test/**/*@(spec.js)\"",
-    "test:raw": "webpack --config webpack.config.js --env ENTRY_DIRS=entries --env CHUNK_DIRS=sharedComps,redux --env ROOT=$(pwd)/../.. --env OVERRIDE_COMPONENT_WEBPACK=test/webpack.config.symlinks.override.js --env REACT4XP_CONFIG_FILE=build/constants.json --env VERBOSE=true ",
-    "test:override": "webpack --config webpack.config.js --env OVERRIDE_COMPONENT_WEBPACK=test/webpack.config.override.js --env ROOT=$(pwd)/../.. --env VERBOSE=true --env REACT4XP_CONFIG_FILE=build/constants.json ",
-    "babel-min": "babel src/entriesandchunks.js -o entriesandchunks.js --no-comments --minified --compact true",
-    "babel": "babel src/entriesandchunks.js -o entriesandchunks.js --no-comments",
-    "lint": "eslint src/*.js --fix",
-    "postversion": "cd ../.. && ./copytag.js buildcomponents"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+ssh://git@github.com:enonic/react4xp-npm.git"
-  },
-  "author": "Espen Norderud",
-  "license": "Apache-2.0",
-  "bugs": {
-    "url": "https://github.com/enonic/lib-react4xp/issues"
-  },
-  "homepage": "https://github.com/enonic/react4xp-build-components",
-  "peerDependencies": {
-    "@babel/plugin-proposal-object-rest-spread": "^7",
-    "@babel/plugin-transform-arrow-functions": "^7",
-    "@babel/preset-env": "^7",
-    "@babel/preset-react": "^7",
-    "prop-types": "*",
-    "react": "*",
-    "react4xp": "^2",
-    "react-dom": "*",
-    "webpack": "^5",
-    "webpack-cli": "^4",
-    "glob": "^7.1",
-    "stats-webpack-plugin": "^0.7",
-    "core-js": "^3"
-  },
-  "lint-staged": {
-    "src/webpack.config.js": [
-      "eslint --fix",
-      "git add"
-    ]
-  },
-  "directories": {
-    "lib": "lib",
-    "test": "__tests__"
-  }
-}
