+++ packages/react4xp/.babelrc
