+++ packages/react4xp/.eslintrc
