+++ packages/react4xp/buildComponents/webpack.config.js
