+++ packages/react4xp/react4xp.gradle
@@ -116,7 +116,7 @@ task config_tasks {
   doLast {
     // Read the file content into an object
     def REACT4XP_TASKS = [
-      "react4xp_components",
+      "react4xpComponents",
       "react4xp_externals",
       "react4xp_client",
       "react4xp_nashornpolyfills"
@@ -141,14 +141,14 @@ config_tasks.dependsOn += 'config_react4xp'
 
 
 // Compile:
-task react4xp_components(type: NodeTask) {
+task react4xpComponents(type: NodeTask) {
   group 'React4xp'
   description 'Compile the parent projects react components'
 
-  // 1 MANDATORY STEP: react4xp-build-components compiles the components added in this project into runnable/renderable components. See react4xp-build-components docs.
+  // 1 MANDATORY STEP: react4xp/buildComponents compiles the components added in this project into runnable/renderable components. See react4xp/buildComponents docs.
   script = file('node_modules/webpack/bin/webpack.js')
   args = [
-    '--config', 'node_modules/react4xp-build-components/webpack.config.js', // react4xp-build-components comes with react4xp@^0.2.0
+    '--config', 'node_modules/react4xp/buildComponents/webpack.config.js',
     '--color',
     '--env', 'VERBOSE=' + project.ext.react4xp.verbose,
     '--env', 'ENTRY_DIRS=' + project.ext.react4xp.entryDirs,
@@ -169,22 +169,22 @@ task react4xp_components(type: NodeTask) {
   args.addAll(['--env', 'REACT4XP_CONFIG_FILE=' + project.ext.react4xp.masterConfigFileName])
 /*
   if (project.ext.react4xp.verbose) {
-    println "react4xp_components task - args:\n"
+    println "react4xpComponents task - args:\n"
   }*/
 
   inputs.file(project.ext.react4xp.outputFileName)
   inputs.file("package.json")
   inputs.file("package-lock.json")
 }
-react4xp_components.dependsOn += 'config_tasks'
-jar.dependsOn += "react4xp_components"
+react4xpComponents.dependsOn += 'config_tasks'
+jar.dependsOn += "react4xpComponents"
 
 
 task react4xp_externals(type: NodeTask) {
   group 'React4xp'
   description 'Compile the externals (react and react-dom'
 
-  // 1 MANDATORY STEP: react4xp-build-components compiles the components added in this project into runnable/renderable components. See react4xp-build-components docs.
+  // 1 MANDATORY STEP: react4xp/buildComponents compiles the components added in this project into runnable/renderable components. See react4xp/buildComponents docs.
   script = file('node_modules/webpack/bin/webpack.js')
   args = [
     '--config', 'node_modules/react4xp-runtime-externals/webpack.config.js',  // react4xp-runtime-externals comes with react4xp@^0.2.0
@@ -222,7 +222,7 @@ task react4xp_client(type: NodeTask) {
   group 'React4xp'
   description 'Compile the react4xp runtime client'
 
-  // 1 MANDATORY STEP: react4xp-build-components compiles the components added in this project into runnable/renderable components. See react4xp-build-components docs.
+  // 1 MANDATORY STEP: react4xp/buildComponents compiles the components added in this project into runnable/renderable components. See react4xp/buildComponents docs.
   script = file('node_modules/webpack/bin/webpack.js')
   args = [
     '--config', 'node_modules/react4xp-runtime-client/webpack.config.js',   // react4xp-runtime-client comes with react4xp@^0.2.0
