+++ packages/react4xp/react4xp.properties
@@ -2,7 +2,7 @@
 #
 # Copy this to your local react4xp project root to use it, change or comment in/out values if needed.
 #
-# NOTE: These settings need to be read and used by react4xp-buildcomponents to build a constants file under /build/. In this package
+# NOTE: These settings need to be read and used by react4xp/buildComponents to build a constants file under /build/. In this package
 #       (react4xp), there is a react4xp.gradle that takes care of that. Without those steps these settings won't have any effect.
 #
 #       What to do? One of these:
@@ -127,11 +127,11 @@ entryDirs =
 
 ####################################   WEBPACK:
 
-# A minimal webpack config is included with react4xp, to build your react components and their dependencies: See node_modules/react4xp-build-components/webpack.config.js.
+# A minimal webpack config is included with react4xp, to build your react components and their dependencies: See node_modules/react4xp/buildComponents/webpack.config.js.
 # To change this setup, or override or extend that webpack.configjs:
 # make a custom file that default-exports a (env, config) => webpackConfigObject function, where:
 #   - env is the collection of "--env" CLI arguments, and
-#   - config is the default config from react4xp-build-components/webpack.config.js.
+#   - config is the default config from react4xp/buildComponents/webpack.config.js.
 # Basically, manipulate or replace the config object and return it.
 # Example file:
 #             module.exports = function(env, config) {
