+++ packages/react4xp/src/buildComponents/entriesandchunks.js
