+++ packages/react4xp/test/buildComponents/dummy-src/react4xp/_entries/thisIsAnEntry.jsx
