+++ packages/react4xp/test/buildComponents/dummy-src/site/parts/client/client.jsx
