+++ packages/react4xp/test/buildComponents/dummy-src/site/parts/example/example.jsx
