+++ packages/react4xp/test/buildComponents/main/resources/entries/Greeter.jsx
