+++ packages/react4xp/test/buildComponents/main/resources/entries/SimpleGreeter.jsx
