+++ packages/react4xp/test/buildComponents/main/resources/react4xp/sharedComps/ReduxWorldGreeter.jsx
