+++ packages/react4xp/test/buildComponents/main/resources/redux/index.es6
