+++ packages/react4xp/test/buildComponents/main/resources/redux/reducers/greetingsRed.es6
