+++ packages/react4xp/test/buildComponents/main/resources/redux/reducers/index.es6
