+++ packages/react4xp/test/buildComponents/main/resources/site/Greeter.jsx
